backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
firebase-admin==6.4.0
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
google-generativeai==0.8.3
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
openai==1.51.2
scikit-learn==1.4.2
sentence-transformers==2.7.0
spacy==3.7.4
keybert==0.8.4
numpy==1.26.4
scipy==1.11.4
slowapi==0.1.9
python-dotenv==1.2.1
python-json-logger==4.0.0